RIO-5979 (OMX spec 1.1.2. requires OMX timestamps to be expressed in microseconds (ticks))

+++ codecs_v2/audio/gsm_amr/amr_nb/enc/include/gsmamr_encoder_wrapper.h
@@ -49,10 +49,10 @@ class TInputAudioStream
         int32 iMode;
 
         // Start time when samples were taken
-        uint32 iStartTime;
+        uint64 iStartTime;
 
         // End time when samples were taken
-        uint32 iStopTime;
+        uint64 iStopTime;
 };
